+++ src/pages/admin/Admin.vue
@@ -165,7 +165,7 @@
 
 <script setup lang="ts">
 import { ref, computed, onMounted } from "vue";
-import LayoutWrapper from "../layouts/LayoutWrapper.vue";
+import LayoutWrapper from "@/layouts/LayoutWrapper.vue";
 import DataTable from "@/components/common/DataTable.vue";
 import {
   emailValidator,
