+++ src/pages/home/Home.vue
@@ -11,6 +11,6 @@
 
 <script setup lang="ts">
 // import { supabase } from '../lib/supabase';
-import LayoutWrapper from '../layouts/LayoutWrapper.vue';
+import LayoutWrapper from '@/layouts/LayoutWrapper.vue';
 
 </script>
