+++ src/pages/home/Profiles.vue
@@ -10,7 +10,7 @@
   
   <script setup lang="ts">
   // import { supabase } from '../lib/supabase';
-  import LayoutWrapper from '../layouts/LayoutWrapper.vue';
+  import LayoutWrapper from '@/layouts/LayoutWrapper.vue';
   
   </script>
